Move implementation to HealthKitMetric
