+++ BeeSwift/HeathKit/HealthKitConfig.swift
@@ -14,27 +14,27 @@ class HealthKitConfig : NSObject {
     
     var metrics : [HealthKitMetric] {
         return [
-            QuantityHealthKitMetric.init(humanText: "Steps", databaseString: "steps", hkIdentifier: HKQuantityTypeIdentifier.stepCount),
-            QuantityHealthKitMetric.init(humanText: "Active energy", databaseString: "activeEnergy", hkIdentifier: HKQuantityTypeIdentifier.activeEnergyBurned),
-            QuantityHealthKitMetric.init(humanText: "Exercise time", databaseString: "exerciseTime", hkIdentifier: HKQuantityTypeIdentifier.appleExerciseTime),
-            QuantityHealthKitMetric.init(humanText: "Weight", databaseString: "weight", hkIdentifier: HKQuantityTypeIdentifier.bodyMass),
-            QuantityHealthKitMetric.init(humanText: "Cycling distance", databaseString: "cyclingDistance", hkIdentifier: HKQuantityTypeIdentifier.distanceCycling),
-            QuantityHealthKitMetric.init(humanText: "Walking/running distance", databaseString: "walkRunDistance", hkIdentifier: HKQuantityTypeIdentifier.distanceWalkingRunning),
-            QuantityHealthKitMetric.init(humanText: "Nike Fuel", databaseString: "nikeFuel", hkIdentifier: HKQuantityTypeIdentifier.nikeFuel),
-            QuantityHealthKitMetric.init(humanText: "Water", databaseString: "water", hkIdentifier: HKQuantityTypeIdentifier.dietaryWater),
+            QuantityHealthKitMetric.init(humanText: "Steps", databaseString: "steps", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.stepCount),
+            QuantityHealthKitMetric.init(humanText: "Active energy", databaseString: "activeEnergy", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.activeEnergyBurned),
+            QuantityHealthKitMetric.init(humanText: "Exercise time", databaseString: "exerciseTime", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.appleExerciseTime),
+            QuantityHealthKitMetric.init(humanText: "Weight", databaseString: "weight", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.bodyMass),
+            QuantityHealthKitMetric.init(humanText: "Cycling distance", databaseString: "cyclingDistance", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.distanceCycling),
+            QuantityHealthKitMetric.init(humanText: "Walking/running distance", databaseString: "walkRunDistance", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.distanceWalkingRunning),
+            QuantityHealthKitMetric.init(humanText: "Nike Fuel", databaseString: "nikeFuel", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.nikeFuel),
+            QuantityHealthKitMetric.init(humanText: "Water", databaseString: "water", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryWater),
             CategoryHealthKitMetric.init(humanText: "Time in bed", databaseString: "timeInBed", hkCategoryTypeIdentifier: HKCategoryTypeIdentifier.sleepAnalysis),
             CategoryHealthKitMetric.init(humanText: "Time asleep", databaseString: "timeAsleep", hkCategoryTypeIdentifier: HKCategoryTypeIdentifier.sleepAnalysis),
 //            CategoryHealthKitMetric.init(humanText: "Stand hours", databaseString: "standHour", hkCategoryTypeIdentifier: HKCategoryTypeIdentifier.appleStandHour),
-            QuantityHealthKitMetric.init(humanText: "Resting energy", databaseString: "basalEnergy", hkIdentifier: HKQuantityTypeIdentifier.basalEnergyBurned),
-            QuantityHealthKitMetric.init(humanText: "Dietary energy", databaseString: "dietaryEnergy", hkIdentifier: HKQuantityTypeIdentifier.dietaryEnergyConsumed),
-            QuantityHealthKitMetric.init(humanText: "Dietary protein", databaseString: "dietaryProtein", hkIdentifier: HKQuantityTypeIdentifier.dietaryProtein),
-            QuantityHealthKitMetric.init(humanText: "Dietary sugar", databaseString: "dietarySugar", hkIdentifier: HKQuantityTypeIdentifier.dietarySugar),
-            QuantityHealthKitMetric.init(humanText: "Dietary carbs", databaseString: "dietaryCarbs", hkIdentifier: HKQuantityTypeIdentifier.dietaryCarbohydrates),
-            QuantityHealthKitMetric.init(humanText: "Dietary fat", databaseString: "dietaryFat", hkIdentifier: HKQuantityTypeIdentifier.dietaryFatTotal),
-            QuantityHealthKitMetric.init(humanText: "Dietary saturated fat", databaseString: "dietarySaturatedFat", hkIdentifier: HKQuantityTypeIdentifier.dietaryFatSaturated),
-            QuantityHealthKitMetric.init(humanText: "Dietary sodium", databaseString: "dietarySodium", hkIdentifier: HKQuantityTypeIdentifier.dietarySodium),
-            QuantityHealthKitMetric.init(humanText: "Swimming strokes", databaseString: "swimStrokes", hkIdentifier: HKQuantityTypeIdentifier.swimmingStrokeCount),
-            QuantityHealthKitMetric.init(humanText: "Swimming distance", databaseString: "swimDistance", hkIdentifier: HKQuantityTypeIdentifier.distanceSwimming),
+            QuantityHealthKitMetric.init(humanText: "Resting energy", databaseString: "basalEnergy", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.basalEnergyBurned),
+            QuantityHealthKitMetric.init(humanText: "Dietary energy", databaseString: "dietaryEnergy", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryEnergyConsumed),
+            QuantityHealthKitMetric.init(humanText: "Dietary protein", databaseString: "dietaryProtein", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryProtein),
+            QuantityHealthKitMetric.init(humanText: "Dietary sugar", databaseString: "dietarySugar", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietarySugar),
+            QuantityHealthKitMetric.init(humanText: "Dietary carbs", databaseString: "dietaryCarbs", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryCarbohydrates),
+            QuantityHealthKitMetric.init(humanText: "Dietary fat", databaseString: "dietaryFat", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryFatTotal),
+            QuantityHealthKitMetric.init(humanText: "Dietary saturated fat", databaseString: "dietarySaturatedFat", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietaryFatSaturated),
+            QuantityHealthKitMetric.init(humanText: "Dietary sodium", databaseString: "dietarySodium", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.dietarySodium),
+            QuantityHealthKitMetric.init(humanText: "Swimming strokes", databaseString: "swimStrokes", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.swimmingStrokeCount),
+            QuantityHealthKitMetric.init(humanText: "Swimming distance", databaseString: "swimDistance", hkQuantityTypeIdentifier: HKQuantityTypeIdentifier.distanceSwimming),
             CategoryHealthKitMetric.init(humanText: "Mindful minutes", databaseString: "mindfulMinutes", hkCategoryTypeIdentifier: HKCategoryTypeIdentifier.mindfulSession)
         ]
     }
